+++ python/edgar_filing_searcher/models.py
@@ -1,6 +1,8 @@
 """Create table models for database"""
 import hashlib
 from dataclasses import dataclass
+from datetime import date
+
 from edgar_filing_searcher.database import db
 
 
@@ -26,7 +28,7 @@ class EdgarFiling(db.Model):
     """Define EdgarFiling Table"""
     accession_no: str
     cik_no: str
-    filing_date: db.Date
+    filing_date: date
 
     __tablename__ = 'edgar_filing'
     accession_no = db.Column(db.String, primary_key=True)
@@ -66,30 +68,31 @@ class Data13f(db.Model):
                "ssh_prnamt='%s', ssh_prnamt_type='%s', putCall='%s', " \
                "investment_discretion='%s', other_manager='%s', voting_authority_sole='%s', " \
                "voting_authority_shared='%s', voting_authority_none='%s')>" % (
-            self.accession_no, self.cik_no, self.name_of_issuer,
-            self.title_of_class, self.cusip, self.value,
-            self.ssh_prnamt, self.ssh_prnamt_type, self.putCall,
-            self.investment_discretion, self.other_manager, self.voting_authority_sole,
-            self.voting_authority_shared, self.voting_authority_none)
+                   self.accession_no, self.cik_no, self.name_of_issuer,
+                   self.title_of_class, self.cusip, self.value,
+                   self.ssh_prnamt, self.ssh_prnamt_type, self.putCall,
+                   self.investment_discretion, self.other_manager, self.voting_authority_sole,
+                   self.voting_authority_shared, self.voting_authority_none)
 
     def create_data_13f_primary_key(self):
         """Uses hash to generate Primary Key based on original row data for Data13f table"""
-        data_13f_row_list = [self.accession_no,
-                              self.cik_no,
-                              self.name_of_issuer,
-                              self.title_of_class,
-                              self.title_of_class,
-                              self.cusip,
-                              self.value,
-                              self.ssh_prnamt,
-                              self.ssh_prnamt_type,
-                              self.put_call,
-                              self.investment_discretion,
-                              self.other_manager,
-                              self.voting_authority_sole,
-                              self.voting_authority_shared,
-                              self.voting_authority_none
-                              ]
+        data_13f_row_list = [
+            self.accession_no,
+            self.cik_no,
+            self.name_of_issuer,
+            self.title_of_class,
+            self.title_of_class,
+            self.cusip,
+            self.value,
+            self.ssh_prnamt,
+            self.ssh_prnamt_type,
+            self.put_call,
+            self.investment_discretion,
+            self.other_manager,
+            self.voting_authority_sole,
+            self.voting_authority_shared,
+            self.voting_authority_none
+        ]
         full_str = ''.join(str(cell) for cell in data_13f_row_list)
         result = hashlib.md5(full_str.encode())
         return result.hexdigest()
